+++ test/test.h
@@ -1,0 +1,23 @@
+//
+// Created by Jonathan Deaton on 3/29/18.
+//
+
+#ifndef LISP_TEST_H
+#define LISP_TEST_H
+
+// Terminal colors
+#define KNRM  "\x1B[0m"
+#define KRED  "\x1B[31m"
+#define KGRN  "\x1B[32m"
+#define KYEL  "\x1B[33m"
+#define KBLU  "\x1B[34m"
+#define KMAG  "\x1B[35m"
+#define KCYN  "\x1B[36m"
+#define KWHT  "\x1B[37m"
+#define RESET "\033[0m"
+
+#define PASS "\x1B[32mPASS\033[0m"
+#define FAIL "\x1B[31mFAIL\033[0m"
+
+
+#endif //LISP_TEST_H
